requirements.txt
asgiref==3.7.2
Django==5.0
Pillow==10.1.0
psycopg2-binary==2.9.9
sqlparse==0.4.4
typing_extensions==4.9.0

django-autoslug~=1.9.9